Move sensor_models to models folder, add std names
