+++ plenario/apiary/__init__.py
@@ -1,14 +1,14 @@
 from collections import defaultdict
-from flask import redirect, render_template
+
+from flask import redirect
 from flask_admin import Admin, AdminIndexView, expose
-from flask_admin.menu import MenuLink, url_for
+from flask_admin.menu import url_for
 from flask_login import current_user
 
-from plenario.database import session
-from plenario.sensor_network.sensor_models import FeatureOfInterest
-from plenario.sensor_network.sensor_models import NetworkMeta, NodeMeta, Sensor
-
 from admin_view import admin_views
+from plenario.database import session
+from plenario.models.SensorNetwork import FeatureMeta
+from plenario.models.SensorNetwork import NetworkMeta, NodeMeta, SensorMeta
 from views import blueprint, redis
 
 
@@ -35,8 +35,8 @@ admin = Admin(
     url="/apiary",
 )
 
-admin.add_view(admin_views["FOI"](FeatureOfInterest, session))
-admin.add_view(admin_views["Sensor"](Sensor, session))
+admin.add_view(admin_views["FOI"](FeatureMeta, session))
+admin.add_view(admin_views["Sensor"](SensorMeta, session))
 admin.add_view(admin_views["Network"](NetworkMeta, session))
 admin.add_view(admin_views["Node"](NodeMeta, session))
