+++ plenario/models/SensorNetwork.py
@@ -38,7 +38,7 @@ class NodeMeta(Base):
     id = Column(String, primary_key=True)
     sensor_network = Column(String, ForeignKey('sensor__network_metadata.name'), primary_key=True)
     location = Column(Geometry(geometry_type='POINT', srid=4326))
-    sensors = relationship('Sensor', secondary='sensor__sensor_to_node')
+    sensors = relationship('SensorMeta', secondary='sensor__sensor_to_node')
     info = Column(JSONB)
 
     column_editable_list = ("sensors", "info")
@@ -64,27 +64,7 @@ class NodeMeta(Base):
         return '<Node "{}">'.format(self.id)
 
 
-class FeatureOfInterest(Base):
-    __tablename__ = 'sensor__features_of_interest'
-
-    name = Column(String, primary_key=True)
-    observed_properties = Column(JSONB)
-
-    @staticmethod
-    def index(network_name=None):
-        features = []
-        for node in session.query(NodeMeta).all():
-            if network_name is None or node.sensor_network.lower() == network_name.lower():
-                for sensor in node.sensors:
-                    for prop in sensor.observed_properties.itervalues():
-                        features.append(prop.split('.')[0].lower())
-        return list(set(features))
-
-    def __repr__(self):
-        return '<Feature "{}">'.format(self.name)
-
-
-class Sensor(Base):
+class SensorMeta(Base):
     __tablename__ = 'sensor__sensors'
 
     name = Column(String, primary_key=True)
@@ -102,3 +82,23 @@ class Sensor(Base):
 
     def __repr__(self):
         return '<Sensor "{}">'.format(self.name)
+
+
+class FeatureMeta(Base):
+    __tablename__ = 'sensor__features_of_interest'
+
+    name = Column(String, primary_key=True)
+    observed_properties = Column(JSONB)
+
+    @staticmethod
+    def index(network_name=None):
+        features = []
+        for node in session.query(NodeMeta).all():
+            if network_name is None or node.sensor_network.lower() == network_name.lower():
+                for sensor in node.sensors:
+                    for prop in sensor.observed_properties.itervalues():
+                        features.append(prop.split('.')[0].lower())
+        return list(set(features))
+
+    def __repr__(self):
+        return '<Feature "{}">'.format(self.name)
