+++ tests/models/test_node_meta.py
@@ -5,7 +5,7 @@ class TestNodeMeta(unittest.TestCase):
 
     def test_nearest_neighbor_with_valid_node(self):
 
-        from plenario.sensor_network.sensor_models import NodeMeta
+        from plenario.models.SensorNetwork import NodeMeta
 
         expected = "NODE_DEV_2"
         observed = NodeMeta.nearest_neighbor_to("NODE_DEV_1")
